v0
The final digit in an ISBN is a check value. It builds redundancy into the number as a verification tool.
This program checks to see if the submitted ISBN (10 or 13) meets the verification criteria. 
If a 9 or 12 digit number is submitted, it offers the fitting check number.
If a handful of other characters are included (space, dash, underscore, etc.), they are removed.