Original Upload
Takes a user input for the name of an interstate and, based on the 'rules' of the naming system, it gives information on that specific interstate.
Not all characters are accounted for, some will break it. But many are removed or stop the script.
It doesn't include any data on which interstates exist, so it doesn't differentiate between imaginary and constructed routes.